Changed number of children type to int, fixed datareader writer to console

+++ ConsoleApp/Models/ImportedObject.cs
@@ -11,7 +11,7 @@
         public string ParentType { get; set; }
         public string DataType { get; set; }
         public string IsNullable { get; set; }
-        public double NumberOfChildren { get; set; }
+        public int NumberOfChildren { get; set; }
 
         public ImportedObject(string[] values)
         {
